refactor: move login and register strategy to each file
