+++ src/auth/register-strategy.ts
@@ -4,36 +4,10 @@ import bcrypt from 'bcrypt';
 import uuid from 'uuid';
 import Users from '../databases/models/users';
 import ErrorMessage from '../error/error-message';
-import { logger } from '../index';
 import config from '../config';
+import { logger } from '../index';
 
-const authLocalStrategy = (): void => {
-  passport.use('login', new LocalStrategy({
-    usernameField: 'email',
-    passwordField: 'password'
-  }, async (email, password, done) => {
-    try {
-      const user = await Users.findOne({
-        where: {
-          email
-        }
-      });
-
-      if (!user) return done(null, false, { message: ErrorMessage.USER_NOT_FOUND });
-
-      const compared = bcrypt.compareSync(password, user.password);
-      if (!compared) return done(null, false, { message: ErrorMessage.USER_NOT_FOUND });
-
-      logger.info(`${user.uuid} ${user.email} 님이 로그인 중입니다.`);
-
-      return done(null, user);
-    } catch (e) {
-      logger.error('로그인 진행 중 오류가 발생하였습니다.');
-      logger.error(e);
-      return done(e);
-    }
-  }));
-
+export default () => {
   passport.use('register', new LocalStrategy({
     usernameField: 'email',
     passwordField: 'password'
@@ -69,5 +43,3 @@ const authLocalStrategy = (): void => {
     }
   }));
 };
-
-export default authLocalStrategy;
